ARM Compile fix.
svn path=/trunk/KDE/kdebase/workspace/; revision=1057628

+++ clients/oxygen/oxygenshadowcache.h
@@ -78,7 +78,7 @@ namespace Oxygen
 
         // even if shadows are disabled,
         // you need a minimum size to allow corner rendering
-        return qMax( size, 5.0 );
+        return qMax( size, qreal(5.0) );
     }
 
     //! get shadow matching client
